Events consider human readable barcodes
The previous regular expression was splitting the barcodes into chunks
of digits, rather than words, which meant human readbale barcodes had
their numeric portion extracted. When this was used in the search no
plates were returned, to audit creation was skipped.

This change splits into chunks of /w which include all characters which
are part of the barcodes.

Whic commit also adds some testing for the delayed job itself, which was
previously missing.